requirements.txt
python-dotenv
flask
slackclient
requests
requests_oauthlib
